Add plan dict to Azure template per cloud_vpn_initiator_azure_image_needs_plan var
Images like RHEL which are free do not need a plan, the API returns a bogus error
if passed in as if the image needs marketplace programmatic access, but reality is
passing the dict is the issue.
Will probably add a task later on to infer this var by querying image facts, for now
user can just pass cloud_vpn_initiator_azure_image_needs_plan var to True work it around.